Fix bug in reading ELECTRONS namelist which goes unnoticed in ifort and pgi, but detected by gfortran and g95. Running by reading from input file is now working, however no checks are being done for inconsistent input and setting options are not working yet.

+++ m_input_vars.f90
@@ -19,7 +19,7 @@ MODULE m_input_vars
   CHARACTER(56) :: KS_Solve_Method
   INTEGER :: icg_beta
   INTEGER :: electron_maxstep
-  INTEGER :: mixing_beta
+  REAL(8) :: mixing_beta
   CHARACTER(56) :: diagonalization
   NAMELIST /ELECTRONS/ KS_Solve_Method, icg_beta, electron_maxstep, mixing_beta, diagonalization
